AirTestPlatform/requirements.txt
APScheduler
pandas
pymysql
cryptography
django-cors-headers
django-excel
openpyxl
requests